+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version: 1.9.0
+
+### New
+
+- New version of `Block` struct with out messages queue updates for foreign workchains
+- Added `GlobalCapabilities::CapWc2WcQueueUpdates`
+- Added new functions related with out messages queue updates
+
 ## Version: 1.8.19
 
 ### Fix